requirements.txt
astroid==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.6
djangorestframework==3.7.1
google-api-python-client-py3==1.2
httplib2==0.10.3
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
py==1.4.34
pylint==1.7.4
pytest==3.2.3
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
wrapt==1.10.11
